updated products and .env

+++ router/products.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 const Product = require('../models/Products');
 
 //Middleware för att verifiera token
-const authenticateToken = require('../middleware/loggedIn'); 
+const authenticateToken = require('../middleware/loggedIn');
 
 // Hämtar alla produkter
 router.get('/products', async (req, res) => {
@@ -12,47 +12,60 @@ router.get('/products', async (req, res) => {
 
 // hämtar specifik produkt efter ID
 router.get('/products/:id', async (req, res) => {
-  const product = await Product.findById({ // Variabel för att hämta enstaka produkt
+  const product = await Product.findById({
+    // Variabel för att hämta enstaka produkt
     _id: req.params.id,
   });
   res.json(product);
 });
 
+//söker upp och uppdaterar en produkt utifrån produkt-id
 router.patch('/products/:id', authenticateToken, async (req, res) => {
-  try {
-    // Variabel för den valda produkten
-    const current = await Product.findById(req.params.id); 
+  //kollar om användaren är admin
+  if (req.user.role === 'admin') {
+    try {
+      // Variabel för den valda produkten
+      const current = await Product.findById(req.params.id);
 
-    // Variabel för att hitta och modifiera vald produkt.
-    const item = await Product.findByIdAndUpdate(
-      { _id: req.params.id },
-      {
-        $set: {
-          // Ändrar till valda värdet eller grundvärdet i produkten.
-          title: req.body.title || current.title,
-          price: req.body.price || current.price,
-          shortDesc: req.body.shortDesc || current.shortDesc,
-          longDesc: req.body.longDesc || current.longDesc,
-        },
-      }
-    );
+      // Variabel för att hitta och modifiera vald produkt.
+      const item = await Product.findByIdAndUpdate(
+        { _id: req.params.id },
+        {
+          $set: {
+            // Ändrar till valda värdet eller grundvärdet i produkten.
+            title: req.body.title || current.title,
+            price: req.body.price || current.price,
+            shortDesc: req.body.shortDesc || current.shortDesc,
+            longDesc: req.body.longDesc || current.longDesc,
+          },
+        }
+      );
 
-    res.json(item);
-  } catch (err) {
-    res.json({ message: err });
+      res.json(item);
+    } catch (err) {
+      res.json({ message: err });
+    }
+  } else {
+    res.send('you are not admin');
   }
 });
 
+// söker upp och raderar en produkt utifrån produkt-id
 router.delete('/products/:id', authenticateToken, async (req, res) => {
-  try {
-    // Variabel för den borttagna produkten
-    const removedItem = await Product.deleteOne({
-      _id: req.params.id,
-    });
+  //kollar om användaren är admin
+  if (req.user.role === 'admin') {
+    try {
+      // Variabel för den borttagna produkten
+      const removedItem = await Product.deleteOne({
+        _id: req.params.id,
+      });
 
-    res.json(removedItem);
-  } catch (err) {
-    res.json({ message: err });
+      res.json(removedItem);
+    } catch (err) {
+      res.json({ message: err });
+    }
+  } else {
+    res.send('you are not admin');
   }
 });
